在模拟器试验层面，完成导入导出功能

+++ app/src/main/java/myapp/chronify/data/nife/Nife.kt
@@ -40,7 +40,7 @@ data class Nife(
     val endDT: LocalDateTime? = null,
     // for cyclical schedule
     val period: PeriodType? = null,
-    val periodMultiple: Short = 1,
+    val periodMultiple: Int = 1,
     val triggerTimes: Set<Int> = emptySet(),
     // extra info
     val description: String = "",
